Allow numeric search string with length < 3 - move common validation method
