+++ src/lib/SearchUtils.js
@@ -1,19 +1,7 @@
-const NOT_VALID_REGEX = /[^a-zA-Z\d.]/g;
-const ADDRESS_PREFIXES = ['zen', 'tzn'];
-const MIN_ADDRESS_SEARCH_LENGTH = 7;
+import {isSearchStringValid as validateSearchString, NOT_VALID_REGEX} from '../common/validations/search';
 
 export default {
-  validateSearchString(searchString) {
-    return (
-      searchString &&
-      searchString.length >= 3 &&
-      !NOT_VALID_REGEX.test(searchString) &&
-      (!ADDRESS_PREFIXES.includes(searchString.substring(0, 3)) ||
-        searchString.length >= MIN_ADDRESS_SEARCH_LENGTH) &&
-      (!ADDRESS_PREFIXES.map(item => `c${item}`).includes(searchString.substring(0, 4)) ||
-        searchString.length >= MIN_ADDRESS_SEARCH_LENGTH + 1)
-    );
-  },
+  validateSearchString,
   isCompleteAddress(searchString) {
     return searchString.indexOf('zen1') === 0 && searchString.length === 63;
   },
